+++ packages/soda-antd/CHANGELOG.md
@@ -1,5 +1,15 @@
 # soda-antd
 
+## 0.8.3
+
+### Patch Changes
+
+- 允许 tree 为空的情况
+- Updated dependencies
+    - deepsea-tools@5.40.2
+    - soda-hooks@6.12.5
+    - soda-react@0.0.20
+
 ## 0.8.2
 
 ### Patch Changes