+++ packages/soda-react-router/CHANGELOG.md
@@ -1,5 +1,12 @@
 # soda-react-router
 
+## 1.1.33
+
+### Patch Changes
+
+- Updated dependencies
+    - soda-hooks@6.12.5
+
 ## 1.1.32
 
 ### Patch Changes